fix typescript error
