+++ src/dialog/overflow-menu/overflow-menu-custom-pane.component.ts
@@ -2,7 +2,7 @@ import { AfterViewInit, Component, ElementRef, Optional } from "@angular/core";
 import { position } from "@carbon/utils-position";
 import { I18n } from "carbon-components-angular/i18n";
 import { ElementService } from "carbon-components-angular/utils";
-
+import { closestAttr } from "carbon-components-angular/utils";
 import { Dialog } from "../dialog.component";
 
 @Component({
@@ -41,7 +41,7 @@ export class OverflowMenuCustomPane extends Dialog implements AfterViewInit {
 			* we also move the element by half of it's own width, since
 			* position service will try and center everything
 			*/
-			const closestRel = this.closestAttr("position", ["relative", "fixed", "absolute"]);
+			const closestRel = closestAttr("position", ["relative", "fixed", "absolute"], this.elementRef.nativeElement);
 			const topFix = closestRel ? closestRel.getBoundingClientRect().top * -1 : 0;
 			const leftFix = closestRel ? closestRel.getBoundingClientRect().left * -1 : 0;
 
@@ -59,24 +59,5 @@ export class OverflowMenuCustomPane extends Dialog implements AfterViewInit {
 		if (!this.dialogConfig.menuLabel) {
 			this.dialogConfig.menuLabel = this.i18n.get().OVERFLOW_MENU.OVERFLOW;
 		}
-	}
-
-	closestAttr(s, t) {
-		let el = this.elementRef.nativeElement;
-
-		do {
-			if (
-				this.matchesAttr(el, s, t)
-			) {
-				return el;
-			}
-			el = el.parentElement || el.parentNode;
-		} while (el !== null && el.nodeType === 1);
-		return null;
-	}
-
-	matchesAttr(el, attr, val) {
-		const styles = window.getComputedStyle(el);
-		return val.includes(styles[attr]);
 	}
 }
